Fix DNT for logged in users.
We were previously, in error, only respecting the DNT header if you were
logged out. For logged in users we were only using the Ello analytics
setting.

With this change we respect both the DNT header and the Ello setting. A
request not to be tracked by either will result in no tracking data.